Initial iteration of the carbon-mesh-viewer
Initial iteration of the carbon mesh viewer

Rendering is done with vulkan, window creation is done with glfw.

Currently the shader path is very hardcoded, need to figure out how to get it working correctly (preferably with a build step).

To view a cmf file you simply drag and drop