Carbon mesh viewer (#9)
* Initial iteration of the carbon-mesh-viewer

Initial iteration of the carbon mesh viewer

Rendering is done with vulkan, window creation is done with glfw.

Currently the shader path is very hardcoded, need to figure out how to get it working correctly (preferably with a build step).

To view a cmf file you simply drag and drop

* Adding shadercache

Shaders are now built into a header file during compilation using glslc and a shader script,

Removed the effect class, built test shaders and glsllangvalidator.exe

* Small fixes

Defaulting to wireframe, fixed the cull mode and made the test color white again

* Refactorings

Refactored the whole thing so now we have the following structure:

Rendering is split into the following sections:
1. Renderer which contains all base rendering structs, device, swapchain etc. This object is passed around to all object creation, initialization and releases.
2. ModelRenderer which renders a model
3. UiRenderer will be created later.

The app state will be stored by the application, which can then send updates to the renderers as needed.

Removed unneeded objects and moved the vulkan folder to rendering

MAde resizing work

* Renaming CmfData to CmfContent

As it was very confusing to have a member variable on the application called m_cmfData which has a member variable which was also called m_cmfData.

So the top level class is called CmfContent

* formatting

* Some commenting and initialization of members

* Camera

Small implementation for the camera

* Added some simple face normal lighting and enabled fill rendering

* Fixing compiler warnings

* Fixing the initial state of the camera

* Fixed CmfContent not correctly converting offsets to pointers.

* First commit of optimizer.

* Fixed dangling pointers and added compression tests.

* Buffer compression in the optimizer

* Fixed Mac compilation issues.

* Update CMakeLists.txt

Making the shader cache be regenerated when ever shaders change

* Fixing the camera

Added smoother rotation
Added panning

* New shaders, track mesh areas, simple LOD visualization.

* Fixed shader cache generator.

* Update camera.cpp

Made the rotation be non existent at the beginning
made the smooth lerping be slightly longer

* Update CMakeLists.txt

Fixing a bug where the generatedShaderCache would not always get generated

* Formatting

---------

Co-authored-by: ccpisheden <isheden@ccpgames.com>
Co-authored-by: ccpisheden <136614133+ccpisheden@users.noreply.github.com>
Co-authored-by: Platform Automation <teamcoreexternal@ccpgames.com>